+++ client/package.json
@@ -8,6 +8,7 @@
     "material-ui-icons": "^1.0.0-beta.36",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
     "yaml-loader": "^0.5.0"
   },
