+++ client/public/manifest.json
@@ -1,21 +1,11 @@
 {
-  "short_name": "mojkredit.si",
+  "short_name": "finster.si",
   "name": "Zbirnik kreditov",
   "icons": [
     {
       "src": "favicon.ico",
       "sizes": "64x64 32x32 24x24 16x16",
       "type": "image/x-icon"
-    },
-    {
-      "src": "logo192.png",
-      "type": "image/png",
-      "sizes": "192x192"
-    },
-    {
-      "src": "logo512.png",
-      "type": "image/png",
-      "sizes": "512x512"
     }
   ],
   "start_url": ".",
